+++ src/main/java/com/ecole/entity/Periode.java
@@ -12,7 +12,7 @@ public class Periode {
     private Long id;
 
     @Column(name = "annee_scolaire_id")
-    private Integer anneeScolaireId;
+    private Long anneeScolaireId;
 
     @Column(nullable = false, length = 100)
     private String libelle;
@@ -45,11 +45,11 @@ public class Periode {
         this.id = id;
     }
 
-    public Integer getAnneeScolaireId() {
+    public Long getAnneeScolaireId() {
         return anneeScolaireId;
     }
 
-    public void setAnneeScolaireId(Integer anneeScolaireId) {
+    public void setAnneeScolaireId(Long anneeScolaireId) {
         this.anneeScolaireId = anneeScolaireId;
     }
